requirements.txt
# Core dependencies
streamlit>=1.22.0
requests>=2.28.2
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
cachetools>=5.3.0

# Testing
pytest>=7.3.1